User acceptance testing #1 (bugs/syntax/logic errors)
• File name checking in case workflow
• More efficient polling method
• Enriched report generation
• Starting message when multi tool begins
• Scoring metric rounded to reasonable figure
• Extra logic on virustotal results to produce cleaner results
• Various syntax fixes
• Output message highlighting screenshot path for URLscan
• Formatting issues fixes
• Logic errors (vpn & proxy printing everytime in reports) fixed

+++ features/case_workflow_tool.py
@@ -23,6 +23,9 @@ def file_creation(investigation_path):
         if not filename.endswith(".txt"):
             filename += ".txt"
         report_path = os.path.join(investigation_path, filename)
+        if os.path.exists(report_path):
+            print(f"A file named '{filename}' already exists. Please choose a different filename.")
+            continue
         try:
             # Varalyze file template
             with open(report_path, "w") as file:
